msm8916-common: Disable Battery LED
AOSiP Battery light customizations no longer have the option to disable the charging light.
On Moto 8916 devices, the charging light is very intrusive so we will permanently disable it in the lights HAL.

Also, use the new AOSiP overlay setting to hide the BatteryLED settings.
msm8916-common: Disable Battery LED

AOSiP Battery light customizations no longer have the option to disable the charging light.
On Moto 8916 devices, the charging light is very intrusive so we will permanently disable it in the lights HAL.